requirements.txt
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
pytesseract==0.3.9
python-dotenv==0.19.2
